Update Checkbox.jsx
The checkbox tabIndex should be a number and I think should match the antd component shown here:

https://github.com/ant-design/ant-design/blob/3ffc7f6ac207f45f1743b1b943dd24586b91152b/components/checkbox/Checkbox.tsx#L21

But after starting the pull request I noticed this was already done and suggested it be changed this number or string. Perhaps merge one of them as currently it is wrong and might save more people coming here to fix it. Cheers :)